Initial commit
Initial version of the project which fulfills the basic requirements of
collecting data from a Varia sensor, summarizing it, and storing the
result in a JSON file.

Known issues:
- Data from the sensor appears to read 10-15% higher than a reference
device. This is likely an issue in the upstream library and needs to be
investigated further. GIGO.
- Application requires two keyboard interrupts to shut down. This is
not a major issue, but can be improved.

Todos:
- Add unit tests
- Add data viewer