Add calibration and thresholds
After testing with a reference vehicle, data from the sensor was found
to be reading high. Unsure if this is just mine, or if they all behave
the save way. Added ability to apply calibration slope and offset to
allow individual sensors to be calibrated.

In testing, also found that the quality of the data drops off below
about 8 metres. Added upper and lower threshold bounds to keep only
the best data.

Store the new parameters in the config file so they don't need to be
specified every time.